requirements.txt
twitter-api-client>=0.10.11
schedule>=1.1.0
pymongo>=4.6.0